+++ common/src/main/java/mca/network/GetVillagerRequest.java
@@ -1,8 +1,13 @@
 package mca.network;
 
+import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
 import mca.cobalt.network.Message;
 import mca.cobalt.network.NetworkHandler;
+import mca.entity.ai.relationship.family.FamilyTree;
+import mca.entity.ai.relationship.family.FamilyTreeNode;
 import mca.network.client.GetVillagerResponse;
 import mca.server.world.data.PlayerSaveData;
 import net.minecraft.entity.Entity;
@@ -12,6 +17,7 @@ import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ServerWorld;
+import net.minecraft.util.Util;
 
 public class GetVillagerRequest implements Message {
     private static final long serialVersionUID = -4415670234855916259L;
@@ -32,15 +38,35 @@ public class GetVillagerRequest implements Message {
         }
     }
 
+    private static void storeNode(NbtCompound data, Optional<FamilyTreeNode> entry, String prefix) {
+        if (entry.isPresent()) {
+            data.putString("tree_" + prefix + "_name", entry.get().getName());
+            data.putUuid("tree_" + prefix + "_uuid", entry.get().id());
+        } else {
+            data.putString("tree_" + prefix + "_name", "");
+            data.putUuid("tree_" + prefix + "_uuid", Util.NIL_UUID);
+        }
+    }
+
     public static NbtCompound getVillagerData(Entity e) {
+        NbtCompound data;
+
         if (e instanceof PlayerEntity) {
-            return PlayerSaveData.get((ServerWorld)e.world, e.getUuid()).getEntityData();
+            data = PlayerSaveData.get((ServerWorld)e.world, e.getUuid()).getEntityData();
         } else if (e instanceof LivingEntity) {
-            NbtCompound villagerData = new NbtCompound();
-            ((MobEntity)e).writeCustomDataToNbt(villagerData);
-            return villagerData;
+            data = new NbtCompound();
+            ((MobEntity)e).writeCustomDataToNbt(data);
         } else {
             return null;
         }
+
+        FamilyTree tree = FamilyTree.get((ServerWorld)e.world);
+        FamilyTreeNode entry = tree.getOrCreate(e);
+
+        storeNode(data, tree.getOrEmpty(entry.spouse()), "spouse");
+        storeNode(data, tree.getOrEmpty(entry.father()), "father");
+        storeNode(data, tree.getOrEmpty(entry.mother()), "mother");
+
+        return data;
     }
 }